request pets vaccine

+++ app/Http/Controllers/API/PetVaccine/PetVaccineController.php
@@ -8,7 +8,10 @@ use App\Http\Requests\PetVaccine\StorePetVaccineRequest;
 use App\Http\Requests\PetVaccine\UpdatePetVaccineRequest;
 use App\Http\Requests\PetVaccine\PartialUpdatePetVaccineRequest;
 use App\Models\PetVaccine\PetVaccine;
+use App\Models\Pet\Pet;
 use App\Services\PetVaccine\PetVaccineService;
+use App\Policies\AccessPetPolicy;
+use App\Policies\AccessPetVaccinePolicy;
 
 class PetVaccineController extends Controller
 {
@@ -38,12 +41,58 @@ class PetVaccineController extends Controller
             return ResponseFormatter::error("Registro no encontrado", 404);
         }
 
-        return ResponseFormatter::success("Registro obtenido exitosamente", 200, $petVaccine);
+        if (!(new AccessPetVaccinePolicy())->access($petVaccine)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para ver esta vacuna de esta mascota',
+                403
+            );
+        }
+
+        $response = $this->service->getById($id);
+
+        if ($response['error']) {
+            return ResponseFormatter::error($response['message'], $response['code']);
+        }
+
+        return ResponseFormatter::success($response['message'], $response['code'], $response['data'] ?? []);
+    }
+
+    public function getVaccinesForPets(string $pet)
+    {
+        $pet = Pet::find($pet);
+
+        if (!$pet) {
+            return ResponseFormatter::error("Registro no encontrado", 404);
+        }
+
+        if (!(new AccessPetPolicy())->access($pet)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para acceder a esta vacunas de esta mascota',
+                403
+            );
+        }
+
+        $response = $this->service->getByPet($pet->id);
+
+        if ($response['error']) {
+            return ResponseFormatter::error($response['message'], $response['code']);
+        }
+
+        return ResponseFormatter::success($response['message'], $response['code'], $response['data'] ?? []);
     }
 
     public function store(StorePetVaccineRequest $request)
     {
-        $response = $this->service->create($request->validated());
+        // Validación de acceso al plan
+        if (!(new AccessPetPolicy())->access($request->pet_id))
+        {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para agregar vacunas a esta mascota',
+                403
+            );
+        }
+        $data = $request->validated();
+        $response = $this->service->create($data);
 
         if ($response['error']) {
             return ResponseFormatter::error($response['message'], $response['code']);
@@ -59,6 +108,16 @@ class PetVaccineController extends Controller
         if (!$petVaccine) {
             return ResponseFormatter::error("Registro no encontrado", 404);
         }
+
+        // Validación de acceso al plan
+        if (!(new AccessPetVaccinePolicy())->access($petVaccine))
+        {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para actualizar las vacunas de esta mascota',
+                403
+            );
+        }
+
 
         $response = $this->service->update($request->validated(), $id);
 
@@ -77,6 +136,15 @@ class PetVaccineController extends Controller
             return ResponseFormatter::error("Registro no encontrado", 404);
         }
 
+        // Validación de acceso al plan
+        if (!(new AccessPetVaccinePolicy())->access($petVaccine))
+        {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para actualizar las vacunas de esta mascota',
+                403
+            );
+        }
+
         $response = $this->service->partialUpdate($request->validated(), $id);
 
         if ($response['error']) {
@@ -92,6 +160,15 @@ class PetVaccineController extends Controller
 
         if (!$petVaccine) {
             return ResponseFormatter::error("Registro no encontrado", 404);
+        }
+
+        // Validación de acceso al plan
+        if (!(new AccessPetVaccinePolicy())->access($petVaccine))
+        {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para eliminar las vacunas de esta mascota',
+                403
+            );
         }
 
         $response = $this->service->delete($id);
